Split index/mod.rs into mod.rs + continuation.rs + phases.rs
mod.rs was 1399 lines and mixed three concerns: CLI arg parsing, the
index() orchestrator, and two largely independent sub-concerns bolted
onto it. Split along those boundaries with no behavior change:

- continuation.rs: the detached-child spawn/handoff machinery used by
  both background-spawn sites (build_detached_child_command,
  spawn_embed_subprocess, redirect_to_background_log, EmbedSpawn,
  detach_embed_eligible, the HANDOFF_* constants).
- phases.rs: the phase-runner entry points (run_phases_3_to_5,
  run_background_phases, run_embed_phases), the embedder-readiness
  wait (wait_for_embedder) it's coupled to, and the embed-skipped
  notice builder (embed_skipped_lines / eprint_embed_skipped_notice).
- mod.rs: IndexArgs and the index() orchestrator only, now calling
  into the two new private submodules.

Tests moved with the code they exercise; the tiny TestCli/clap
fixture is duplicated in continuation.rs's test module rather than
threaded across files. All 141 tests in the module pass unchanged,
cargo clippy -D warnings is clean, and cargo fmt --check is clean.

Follow-up to Johan's review on #780 (spelunk-oss task #304).
